+++ app/src/scenes/phase1/scenes/affected/components/TodoBackpack.js
@@ -32,59 +32,60 @@ export default function TodoBackpack({ lunchBreak }) {
   }
 
   return (
-    <div className="relative bg-white rounded-xl shadow-nina md:shadow-none p-[1rem] md:p-[0rem] m-[1rem] md:m-[0rem] md:w-full overflow-hidden">
+    <div className="mx-[0.5rem] flex overflow-hidden rounded-xl bg-white shadow-nina md:mx-[0rem] md:shadow-none lg:w-2/3">
       <MedicalFileModal isOpen={isMedicalFileModalOpen} onClose={() => setMedicalFileModalOpen(false)} />
-      <div className="max-w-md">
-        <h1 className="text-xl font-bold mb-6 md:pl-16">Dans mon sac...</h1>
-        <div className="text-sm space-y-3 md:pl-16 md:border-l-[1px]">
-          <div className="flex gap-4 items-baseline">
-            <input type="checkbox" name="convocation" id="convocation" checked={todo.convocation} onChange={handleCheck} />
-            <label htmlFor="convocation">
-              Votre{" "}
-              <button onClick={() => setModalConvocationOpen(true)} className="h-6 font-semibold underline-offset-4 underline decoration-2">
-                convocation
-              </button>
-            </label>
-          </div>
 
-          <div className="flex gap-4 items-baseline">
-            <input type="checkbox" name="identite" id="identite" checked={todo.identite} onChange={handleCheck} />
-            <label htmlFor="identite">
-              Votre <strong>pièce d&apos;identité</strong>
-            </label>
-          </div>
+      <div className="relative p-4">
+        <h1 className="mb-6 text-xl font-bold">Dans mon sac...</h1>
 
-          <div className="flex gap-4 items-baseline">
-            <input type="checkbox" name="sanitaire" id="sanitaire" checked={todo.sanitaire} onChange={handleCheck} />
-            <label htmlFor="sanitaire">
-              La{" "}
-              <button onClick={() => setMedicalFileModalOpen(true)} className="h-6 font-semibold underline-offset-4 underline decoration-2">
-                fiche sanitaire
-              </button>{" "}
-              complétée, sous enveloppe destinée au référent sanitaire
-            </label>
-          </div>
-
-          <div className="flex gap-4 items-baseline">
-            <input type="checkbox" name="masques" id="masques" checked={todo.masques} onChange={handleCheck} />
-            <label htmlFor="masques">
-              Deux <strong>masques jetables</strong> à usage médical pour le transport en commun
-            </label>
-          </div>
-
-          {lunchBreak && (
-            <div className="flex gap-4 items-baseline">
-              <input type="checkbox" name="collation" id="collation" checked={todo.collation} onChange={handleCheck} />
-              <label htmlFor="collation">
-                Une <strong>collation ou un déjeuner froid</strong> pour le repas.
-              </label>
-            </div>
-          )}
+        <div className="flex items-baseline gap-4">
+          <input type="checkbox" name="convocation" id="convocation" checked={todo.convocation} onChange={handleCheck} />
+          <label htmlFor="convocation">
+            Votre{" "}
+            <button onClick={() => setModalConvocationOpen(true)} className="h-6 font-semibold underline decoration-2 underline-offset-4">
+              convocation
+            </button>
+          </label>
         </div>
+
+        <div className="flex items-baseline gap-4">
+          <input type="checkbox" name="identite" id="identite" checked={todo.identite} onChange={handleCheck} />
+          <label htmlFor="identite">
+            Votre <strong>pièce d&apos;identité</strong>
+          </label>
+        </div>
+
+        <div className="flex items-baseline gap-4">
+          <input type="checkbox" name="sanitaire" id="sanitaire" checked={todo.sanitaire} onChange={handleCheck} />
+          <label htmlFor="sanitaire">
+            La{" "}
+            <button onClick={() => setMedicalFileModalOpen(true)} className="h-6 font-semibold underline decoration-2 underline-offset-4">
+              fiche sanitaire
+            </button>{" "}
+            complétée, sous enveloppe destinée au référent sanitaire
+          </label>
+        </div>
+
+        <div className="flex items-baseline gap-4">
+          <input type="checkbox" name="masques" id="masques" checked={todo.masques} onChange={handleCheck} />
+          <label htmlFor="masques">
+            Deux <strong>masques jetables</strong> à usage médical pour le transport en commun
+          </label>
+        </div>
+
+        {lunchBreak && (
+          <div className="flex items-baseline gap-4">
+            <input type="checkbox" name="collation" id="collation" checked={todo.collation} onChange={handleCheck} />
+            <label htmlFor="collation">
+              Une <strong>collation ou un déjeuner froid</strong> pour le repas.
+            </label>
+          </div>
+        )}
+        <Arrow className="absolute left-80 top-10 hidden lg:block" />
+        <DontForget className="absolute left-80 top-20 hidden lg:block" />
       </div>
-      <Arrow className="absolute hidden md:block left-80 top-0" />
-      <DontForget className="absolute hidden md:block left-80 top-10" />
-      <SnuBackPack className="block md:hidden xl:block absolute -right-4 md:-right-16 top-4 md:w-64 md:h-64" />
+
+      <SnuBackPack className="mt-4 block flex-none md:hidden md:h-64 md:w-64 xl:block" />
 
       <ModalConvocation open={modalConvocationOpen} setOpen={setModalConvocationOpen} />
     </div>